+++ packages/utils/test/abstract-session.spec.ts
@@ -234,6 +234,27 @@ describe('abstract-session', () => {
       .with.property('code', 'ABORT_ERR')
   })
 
+  it('should not make multiple requests to the only found provider', async function () {
+    const session = new Session()
+    const cid = CID.parse('bafybeifaymukvfkyw6xgh4th7tsctiifr4ea2btoznf46y6b2fnvikdczi')
+    const id = await createEd25519PeerId() // same provider
+
+    session.findNewProviders.callsFake(async function * () {
+      yield {
+        id
+      }
+    })
+    session.queryProvider.callsFake(async () => {
+      // always fails
+      throw new Error('Urk!')
+    })
+
+    await expect(session.retrieve(cid)).to.eventually.be.rejected()
+
+    expect(session.findNewProviders).to.have.property('callCount', 2)
+    expect(session.queryProvider).to.have.property('callCount', 1)
+  })
+
   it('should respect signals', async () => {
     const signal = AbortSignal.timeout(10)
     const session = new Session()