+++ packages/helia/CHANGELOG.md
@@ -1,5 +1,65 @@
 # Changelog
 
+## [4.2.6](https://github.com/ipfs/helia/compare/helia-v4.2.5...helia-v4.2.6) (2024-09-13)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/block-brokers bumped from ^3.0.3 to ^3.0.4
+
+## [4.2.5](https://github.com/ipfs/helia/compare/helia-v4.2.4...helia-v4.2.5) (2024-07-31)
+
+
+### Bug Fixes
+
+* update js-libp2p types ([#570](https://github.com/ipfs/helia/issues/570)) ([b4877b5](https://github.com/ipfs/helia/commit/b4877b5b768895684be90a26f4303ae65fc209e7))
+
+
+### Documentation
+
+* fix grammar - it's -&gt; its ([#565](https://github.com/ipfs/helia/issues/565)) ([155e24d](https://github.com/ipfs/helia/commit/155e24db8c06c33972895d702a656e0c2996f3d9))
+
+
+### Dependencies
+
+* bump aegir from 42.2.11 to 43.0.1 ([#552](https://github.com/ipfs/helia/issues/552)) ([74ccc92](https://github.com/ipfs/helia/commit/74ccc92793a6d0bb4bee714d9fe4fa4183aa4ee8))
+* bump aegir from 43.0.3 to 44.0.1 ([#569](https://github.com/ipfs/helia/issues/569)) ([6952f05](https://github.com/ipfs/helia/commit/6952f05357844e5aa3dffb2afaf261df06b9b7c1))
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/block-brokers bumped from ^3.0.2 to ^3.0.3
+    * @helia/interface bumped from ^4.3.0 to ^4.3.1
+    * @helia/routers bumped from ^1.1.0 to ^1.1.1
+    * @helia/utils bumped from ^0.3.2 to ^0.3.3
+
+## [4.2.4](https://github.com/ipfs/helia/compare/helia-v4.2.3...helia-v4.2.4) (2024-06-18)
+
+
+### Bug Fixes
+
+* update circuit relay server args ([#561](https://github.com/ipfs/helia/issues/561)) ([3577d3d](https://github.com/ipfs/helia/commit/3577d3d106e255ff0d2a1d47a197f04632b903ec))
+
+## [4.2.3](https://github.com/ipfs/helia/compare/helia-v4.2.2...helia-v4.2.3) (2024-05-27)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/block-brokers bumped from ^3.0.1 to ^3.0.2
+    * @helia/utils bumped from ^0.3.1 to ^0.3.2
+
+## [4.2.2](https://github.com/ipfs/helia/compare/helia-v4.2.1...helia-v4.2.2) (2024-05-20)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/block-brokers bumped from ^3.0.0 to ^3.0.1
+    * @helia/utils bumped from ^0.3.0 to ^0.3.1
+
 ## [4.2.1](https://github.com/ipfs/helia/compare/helia-v4.2.0...helia-v4.2.1) (2024-05-02)