+++ packages/http/CHANGELOG.md
@@ -1,5 +1,54 @@
 # Changelog
 
+## [1.0.11](https://github.com/ipfs/helia/compare/http-v1.0.10...http-v1.0.11) (2024-09-13)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/block-brokers bumped from ^3.0.3 to ^3.0.4
+
+## [1.0.10](https://github.com/ipfs/helia/compare/http-v1.0.9...http-v1.0.10) (2024-07-31)
+
+
+### Documentation
+
+* fix grammar - it's -&gt; its ([#565](https://github.com/ipfs/helia/issues/565)) ([155e24d](https://github.com/ipfs/helia/commit/155e24db8c06c33972895d702a656e0c2996f3d9))
+
+
+### Dependencies
+
+* bump aegir from 42.2.11 to 43.0.1 ([#552](https://github.com/ipfs/helia/issues/552)) ([74ccc92](https://github.com/ipfs/helia/commit/74ccc92793a6d0bb4bee714d9fe4fa4183aa4ee8))
+* bump aegir from 43.0.3 to 44.0.1 ([#569](https://github.com/ipfs/helia/issues/569)) ([6952f05](https://github.com/ipfs/helia/commit/6952f05357844e5aa3dffb2afaf261df06b9b7c1))
+* **dev:** bump sinon from 17.0.2 to 18.0.0 ([#536](https://github.com/ipfs/helia/issues/536)) ([62f77df](https://github.com/ipfs/helia/commit/62f77dfbff94a64e9c248f5be54055c18a6427f7))
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/block-brokers bumped from ^3.0.2 to ^3.0.3
+    * @helia/interface bumped from ^4.3.0 to ^4.3.1
+    * @helia/routers bumped from ^1.1.0 to ^1.1.1
+    * @helia/utils bumped from ^0.3.2 to ^0.3.3
+
+## [1.0.9](https://github.com/ipfs/helia/compare/http-v1.0.8...http-v1.0.9) (2024-05-27)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/block-brokers bumped from ^3.0.1 to ^3.0.2
+    * @helia/utils bumped from ^0.3.1 to ^0.3.2
+
+## [1.0.8](https://github.com/ipfs/helia/compare/http-v1.0.7...http-v1.0.8) (2024-05-20)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/block-brokers bumped from ^3.0.0 to ^3.0.1
+    * @helia/utils bumped from ^0.3.0 to ^0.3.1
+
 ## [1.0.7](https://github.com/ipfs/helia/compare/http-v1.0.6...http-v1.0.7) (2024-05-02)