+++ packages/ipns/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [7.2.3](https://github.com/ipfs/helia/compare/ipns-v7.2.2...ipns-v7.2.3) (2024-07-31)
+
+
+### Documentation
+
+* fix grammar - it's -&gt; its ([#565](https://github.com/ipfs/helia/issues/565)) ([155e24d](https://github.com/ipfs/helia/commit/155e24db8c06c33972895d702a656e0c2996f3d9))
+
+
+### Dependencies
+
+* bump aegir from 42.2.11 to 43.0.1 ([#552](https://github.com/ipfs/helia/issues/552)) ([74ccc92](https://github.com/ipfs/helia/commit/74ccc92793a6d0bb4bee714d9fe4fa4183aa4ee8))
+* bump aegir from 43.0.3 to 44.0.1 ([#569](https://github.com/ipfs/helia/issues/569)) ([6952f05](https://github.com/ipfs/helia/commit/6952f05357844e5aa3dffb2afaf261df06b9b7c1))
+* **dev:** bump sinon from 17.0.2 to 18.0.0 ([#536](https://github.com/ipfs/helia/issues/536)) ([62f77df](https://github.com/ipfs/helia/commit/62f77dfbff94a64e9c248f5be54055c18a6427f7))
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^4.3.0 to ^4.3.1
+
 ## [7.2.2](https://github.com/ipfs/helia/compare/ipns-v7.2.1...ipns-v7.2.2) (2024-05-01)